refactor: make mminf/utils and store ipc/flashinfer/profiler code there
