+++ mminf/engine/enc_dec_engine.py
@@ -1,7 +1,7 @@
 import torch
 
 from mminf.engine.base import BaseEngine, EngineType, NodeBatch, NodeOutput
-from mminf.profiler import range_pop, range_push
+from mminf.utils.profiler import range_pop, range_push
 
 
 class EncoderDecoderEngine(BaseEngine):
