+++ mminf/utils/ipc_format.py
